Move ApparentHorizons to ParallelAlgorithms/ApparentHorizonFinder
