+++ src/ParallelAlgorithms/ApparentHorizonFinder/InterpolationTarget.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "ApparentHorizons/InterpolationTarget.hpp"
+#include "ParallelAlgorithms/ApparentHorizonFinder/InterpolationTarget.hpp"
 
 #include <algorithm>
