+++ tests/Unit/CMakeLists.txt
@@ -8,7 +8,6 @@ set(SPECTRE_TESTS)
 set_property(GLOBAL PROPERTY SPECTRE_TESTS_LIB_FUNCTIONS_PROPERTY "")
 set_property(GLOBAL PROPERTY SPECTRE_TESTS_LIBS_PROPERTY "")
 
-add_subdirectory(ApparentHorizons)
 add_subdirectory(ControlSystem)
 add_subdirectory(DataStructures)
 add_subdirectory(Domain)
