+++ tests/Unit/ParallelAlgorithms/CMakeLists.txt
@@ -2,6 +2,7 @@
 # See LICENSE.txt for details.
 
 add_subdirectory(Actions)
+add_subdirectory(ApparentHorizonFinder)
 add_subdirectory(Amr)
 add_subdirectory(Events)
 add_subdirectory(EventsAndTriggers)
